plant_disease_diagnosis/requirements.txt
streamlit~=1.41.1
watchdog
pillow~=11.0.0
google-generativeai
load_dotenv